requirements.txt
python-dotenv
pydub
sounddevice
simpleaudio
gradio==3.42.0
gradio_client==0.5.0
scipy
















